+++ corehq/apps/app_manager/static/app_manager/js/forms/advanced/case_config_ui.js
